Move client from v4 to root
