+++ gspread/client.py
@@ -1,57 +1,45 @@
 # -*- coding: utf-8 -*-
 
 """
-gspread.client
+gspread.v4.client
 ~~~~~~~~~~~~~~
 
 This module contains Client class responsible for communicating with
-Google Data API.
+Google API.
 
 """
-import json
 
-try:
-    import xml.etree.cElementTree as ElementTree
-except:
-    from xml.etree import ElementTree
+import requests
 
-from . import urlencode
-from .ns import _ns
-from .httpsession import HTTPSession
-from .exceptions import RequestError
+from ..utils import finditem
+from ..base import BaseClient
+
+from ..exceptions import SpreadsheetNotFound
+from .exceptions import APIError
 from .models import Spreadsheet
-from .urls import (
-    construct_url,
+
+from ..urls import (
     DRIVE_FILES_API_V2_URL,
     DRIVE_FILES_UPLOAD_API_V2_URL
 )
-from .utils import finditem, extract_id_from_url
-from .exceptions import (SpreadsheetNotFound, UpdateCellError)
-
-from .base import BaseClient
 
 
 class Client(BaseClient):
+    """An instance of this class communicates with Google API.
 
-    """An instance of this class communicates with Google Data API.
+    :param auth: An OAuth2 credential object. Credential objects
+                 are those created by the oauth2client library.
+                 https://github.com/google/oauth2client
+    :param session: (optional) A session object capable of making HTTP requests
+                    while persisting some parameters across requests.
+                    Defaults to `requests.Session <http://docs.python-requests.org/en/master/api/#request-sessions>`_.
 
-    :param auth: An OAuth2 credential object. Credential objects are those created by the
-                 oauth2client library. https://github.com/google/oauth2client
-    :param http_session: (optional) A session object capable of making HTTP requests while persisting headers.
-                                    Defaults to :class:`~gspread.httpsession.HTTPSession`.
-
-    >>> c = gspread.Client(auth=OAuthCredentialObject)
+    >>> c = gspread.v4.Client(auth=OAuthCredentialObject)
 
     """
-    def __init__(self, auth, http_session=None):
+    def __init__(self, auth, session=None):
         self.auth = auth
-        self.session = http_session or HTTPSession()
-
-    def _ensure_xml_header(self, data):
-        if data.startswith(b'<?xml'):
-            return data
-        else:
-            return b'<?xml version="1.0" encoding="utf8"?>' + data
+        self.session = session or requests.Session()
 
     def login(self):
         """Authorize client."""
@@ -62,14 +50,61 @@ class Client(BaseClient):
             http = httplib2.Http()
             self.auth.refresh(http)
 
-        self.session.add_header('Authorization', "Bearer " + self.auth.access_token)
+        self.session.headers.update({
+            'Authorization': 'Bearer %s' % self.auth.access_token
+        })
+
+    def request(
+            self,
+            method,
+            endpoint,
+            params=None,
+            data=None,
+            json=None,
+            files=None,
+            headers=None):
+
+        response = getattr(self.session, method)(
+            endpoint,
+            json=json,
+            params=params,
+            data=data,
+            files=files,
+            headers=headers
+        )
+
+        if response.ok:
+            return response
+        else:
+            raise APIError(response)
+
+    def list_spreadsheet_files(self):
+        files = []
+        page_token = ''
+        url = "https://www.googleapis.com/drive/v3/files"
+        params = {
+            'q': "mimeType='application/vnd.google-apps.spreadsheet'",
+            "pageSize": 1000,
+            'supportsTeamDrives': True,
+            'includeTeamDriveItems': True,
+        }
+
+        while page_token is not None:
+            if page_token:
+                params['pageToken'] = page_token
+
+            res = self.request('get', url, params=params).json()
+            files.extend(res['files'])
+            page_token = res.get('nextPageToken', None)
+
+        return files
 
     def open(self, title):
         """Opens a spreadsheet.
 
         :param title: A title of a spreadsheet.
 
-        :returns: a :class:`~gspread.Spreadsheet` instance.
+        :returns: a :class:`~gspread.v4.models.Spreadsheet` instance.
 
         If there's more than one spreadsheet with same title the first one
         will be opened.
@@ -81,13 +116,17 @@ class Client(BaseClient):
         >>> c.open('My fancy spreadsheet')
 
         """
-        feed = self.get_spreadsheets_feed()
+        try:
+            properties = finditem(
+                lambda x: x['name'] == title,
+                self.list_spreadsheet_files()
+            )
 
-        for elem in feed.findall(_ns('entry')):
-            elem_title = elem.find(_ns('title')).text
-            if elem_title.strip() == title:
-                return Spreadsheet(self, elem)
-        else:
+            # Drive uses different terminology
+            properties['title'] = properties['name']
+
+            return Spreadsheet(self, properties)
+        except StopIteration:
             raise SpreadsheetNotFound
 
     def open_by_key(self, key):
@@ -95,24 +134,13 @@ class Client(BaseClient):
 
         :param key: A key of a spreadsheet as it appears in a URL in a browser.
 
-        :returns: a :class:`~gspread.Spreadsheet` instance.
-
-        :raises gspread.SpreadsheetNotFound: if no spreadsheet with
-                                             specified `key` is found.
+        :returns: a :class:`~gspread.v4.models.Spreadsheet` instance.
 
         >>> c = gspread.authorize(credentials)
         >>> c.open_by_key('0BmgG6nO_6dprdS1MN3d3MkdPa142WFRrdnRRUWl1UFE')
 
         """
-        feed = self.get_spreadsheets_feed()
-        for elem in feed.findall(_ns('entry')):
-            alter_link = finditem(lambda x: x.get('rel') == 'alternate',
-                                  elem.findall(_ns('link')))
-            spreadsheet_id = extract_id_from_url(alter_link.get('href'))
-            if spreadsheet_id == key:
-                return Spreadsheet(self, elem)
-        else:
-            raise SpreadsheetNotFound
+        return Spreadsheet(self, {'id': key})
 
     def openall(self, title=None):
         """Opens all available spreadsheets.
@@ -120,120 +148,22 @@ class Client(BaseClient):
         :param title: (optional) If specified can be used to filter
                       spreadsheets by title.
 
-        :returns: a list of :class:`~gspread.Spreadsheet` instances.
+        :returns: a list of :class:`~gspread.v4.models.Spreadsheet` instances.
 
         """
-        feed = self.get_spreadsheets_feed()
-        result = []
-        for elem in feed.findall(_ns('entry')):
-            if title is not None:
-                elem_title = elem.find(_ns('title')).text
-                if elem_title.strip() != title:
-                    continue
-            result.append(Spreadsheet(self, elem))
+        spreadsheet_files = self.list_spreadsheet_files()
 
-        return result
-
-    def get_spreadsheets_feed(self, visibility='private', projection='full'):
-        url = construct_url('spreadsheets',
-                            visibility=visibility, projection=projection)
-
-        r = self.session.get(url)
-        return ElementTree.fromstring(r.content)
-
-    def get_worksheets_feed(self, spreadsheet,
-                            visibility='private', projection='full'):
-        url = construct_url('worksheets', spreadsheet,
-                            visibility=visibility, projection=projection)
-
-        r = self.session.get(url)
-        return ElementTree.fromstring(r.content)
-
-    def get_cells_feed(self, worksheet,
-                       visibility='private', projection='full', params=None):
-
-        url = construct_url('cells', worksheet,
-                            visibility=visibility, projection=projection)
-
-        if params:
-            params = urlencode(params)
-            url = '%s?%s' % (url, params)
-
-        r = self.session.get(url)
-        return ElementTree.fromstring(r.content)
-
-    def get_feed(self, url):
-        r = self.session.get(url)
-        return ElementTree.fromstring(r.content)
-
-    def del_spreadsheet(self, file_id):
-        """Deletes a spreadsheet.
-
-        :param file_id: a spreadsheet ID (aka file ID.)
-        """
-        url = '{0}/{1}'.format(
-            DRIVE_FILES_API_V2_URL,
-            file_id
-        )
-
-        self.session.delete(url)
-
-    def del_worksheet(self, worksheet):
-        url = construct_url(
-            'worksheet',
-            worksheet,
-            'private',
-            'full',
-            worksheet_version=worksheet.version
-        )
-        self.session.delete(url)
-
-    def get_cells_cell_id_feed(self, worksheet, cell_id,
-                               visibility='private', projection='full'):
-        url = construct_url('cells_cell_id', worksheet, cell_id=cell_id,
-                            visibility=visibility, projection=projection)
-
-        r = self.session.get(url)
-        return ElementTree.fromstring(r.content)
-
-    def put_feed(self, url, data):
-        headers = {'Content-Type': 'application/atom+xml',
-                   'If-Match': '*'}
-        data = self._ensure_xml_header(data)
-
-        try:
-            r = self.session.put(url, data, headers=headers)
-        except RequestError as ex:
-            if ex.args[0] == 403:
-                raise UpdateCellError(ex.args[1])
-            else:
-                raise
-
-        return ElementTree.fromstring(r.content)
-
-    def post_feed(self, url, data):
-        headers = {'Content-Type': 'application/atom+xml'}
-        data = self._ensure_xml_header(data)
-
-        r = self.session.post(url, data, headers=headers)
-
-        return ElementTree.fromstring(r.content)
-
-    def post_cells(self, worksheet, data):
-        headers = {'Content-Type': 'application/atom+xml',
-                   'If-Match': '*'}
-        data = self._ensure_xml_header(data)
-        url = construct_url('cells_batch', worksheet)
-        r = self.session.post(url, data, headers=headers)
-
-        return ElementTree.fromstring(r.content)
+        return [
+            Spreadsheet(self, dict(title=x['name'], **x))
+            for x in spreadsheet_files
+        ]
 
     def create(self, title):
         """Creates a new spreadsheet.
 
         :param title: A title of a new spreadsheet.
 
-        :returns: a :class:`~gspread.Spreadsheet` instance.
+        :returns: a :class:`~gspread.v4.models.Spreadsheet` instance.
 
         .. note::
 
@@ -251,19 +181,29 @@ class Client(BaseClient):
            when you try to create a new spreadsheet.
 
         """
-
-        headers = {'Content-Type': 'application/json'}
-        data = {
+        payload = {
             'title': title,
             'mimeType': 'application/vnd.google-apps.spreadsheet'
         }
-        r = self.session.post(
+        r = self.request(
+            'post',
             DRIVE_FILES_API_V2_URL,
-            json.dumps(data),
-            headers=headers
+            json=payload
         )
         spreadsheet_id = r.json()['id']
         return self.open_by_key(spreadsheet_id)
+
+    def del_spreadsheet(self, file_id):
+        """Deletes a spreadsheet.
+
+        :param file_id: a spreadsheet ID (aka file ID.)
+        """
+        url = '{0}/{1}'.format(
+            DRIVE_FILES_API_V2_URL,
+            file_id
+        )
+
+        self.request('delete', url)
 
     def import_csv(self, file_id, data):
         """Imports data into the first page of the spreadsheet.
@@ -273,7 +213,8 @@ class Client(BaseClient):
         headers = {'Content-Type': 'text/csv'}
         url = '{0}/{1}'.format(DRIVE_FILES_UPLOAD_API_V2_URL, file_id)
 
-        self.session.put(
+        self.request(
+            'put',
             url,
             data=data,
             params={
@@ -289,9 +230,8 @@ class Client(BaseClient):
         :param file_id: a spreadsheet ID (aka file ID.)
         """
         url = '{0}/{1}/permissions'.format(DRIVE_FILES_API_V2_URL, file_id)
-        headers = {'Content-Type': 'application/json'}
 
-        r = self.session.get(url, headers=headers)
+        r = self.request('get', url)
 
         return r.json()['items']
 
@@ -302,8 +242,7 @@ class Client(BaseClient):
         perm_type,
         role,
         notify=True,
-        email_message=None,
-        with_link=False
+        email_message=None
     ):
         """Creates a new permission for a file.
 
@@ -339,25 +278,14 @@ class Client(BaseClient):
                 role='reader'
             )
 
-            # Make the spreadsheet publicly editable by anyone with the link
-
-            gc.insert_permission(
-                '0BmgG6nO_6dprnRRUWl1UFE',
-                None,
-                perm_type='anyone',
-                role='writer',
-                with_link=True
-            )
-
         """
 
         url = '{0}/{1}/permissions'.format(DRIVE_FILES_API_V2_URL, file_id)
 
-        data = {
+        payload = {
             'value': value,
             'type': perm_type,
             'role': role,
-            'withLink': with_link
         }
 
         params = {
@@ -365,13 +293,11 @@ class Client(BaseClient):
             'emailMessage': email_message
         }
 
-        headers = {'Content-Type': 'application/json'}
-
-        self.session.post(
+        self.request(
+            'post',
             url,
-            json.dumps(data),
-            params=params,
-            headers=headers
+            json=payload,
+            params=params
         )
 
     def remove_permission(self, file_id, permission_id):
@@ -385,6 +311,5 @@ class Client(BaseClient):
             file_id,
             permission_id
         )
-        headers = {'Content-Type': 'application/json'}
 
-        self.session.delete(url, headers=headers)
+        self.request('delete', url)
